requirements.txt
setuptools
# Hadron foundation
aistac-foundation
# Data Science packages
pyarrow>=12.0
pandas>=2.0
numpy
scikit-learn
matplotlib
seaborn
scipy
# structure packages
requests
pyyaml
# S3 packages
boto3
botocore
fsspec
s3fs
# others


